moving typedRoutes into serverside lib
