+++ src/shared/lib/routeUtils.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/ban-types */
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { Router } from 'express';
 
 /**
  * These are all the accepted url routes for this application
@@ -149,18 +148,4 @@ export const buildUrl = <P extends AllRoutes>(
   }
 
   return ret;
-};
-
-const router = Router();
-
-export const typedRouter = {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  get: (path: RoutePaths, ...handlers: Array<any>) => {
-    return router.get(path, ...handlers);
-  },
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  post: (path: RoutePaths, ...handlers: Array<any>) => {
-    return router.get(path, ...handlers);
-  },
-  router: router,
 };
